Nexus/NEXUSRJ/backend/requirements.txt
fastapi==0.135.3
sqlalchemy==2.0.41
uvicorn==0.44.0
python-socketio>=5.11.0
python-multipart>=0.0.9
aiohttp>=3.9.0
